FLask test command unable to run on the workflow fix
The command "docker exec -i api-test flask test" is supposed to run unit test on the api container
but this fails because using the flask command we have to export FLASK_APP=<application Entry point> in this case
manage.py

Since neither "RUN export FLASK_APP=manage.py" nor CMD ["export", "FLASK_APP"] configed on the image file  solved the issue
The the only solution i could think of is have a .flaskenv file on the root directory of the application and have FLASK_APP=manage
directive defined.
Since the given directory is not a sensitive information the .flaskenv file can be used to configure basic flask required directives.

Having a python-doenv package installed as a dependency will help consume the directive defined on the .flaskenv file and thus fix the issue of the
application test workflow failing at unit testing stage. ;) HOPEFULY!!!!