User Account Activation
User Account Activation is done using a url sent to the user registered Email address
 instead of sending the email at the registration route. I have opted to hook a callback
to the after_insert sqlalchemy ORM object event.

Once a new user have been registered to the application the send_email would be called

to allow smooth application interaction. the email function runs in a separate thread from
the main application's thread but in sync via the application's context

although the test_mail_runs_in_current_app_context fails, as of the current implementation
because the application context is pushed from a mocked object :)
still tring to think this through tho!!!!!!!!!!!!!!!!