User Account token generation and new token requests
Changes from previous commits

Previously once a user account is created the register route response included both the refresh token and the access token.
I have opted to change the response body of this route to only notify for a successful registration.

The Login route have also been added. The route uses basic authentication:

	on successful login the route responds with both the access token and the refresh token.

	ofcourse the access token have a short life span

to renew the access token the api avails /api/auth/token/renew route
	This route uses token based authentication i.e the refresh token is used to authenticate
	to this route. also othe expired access token is requred to be passed as a payload to the
	request (Supports only post requests)

The application caches both the initial refresh token and access token, since the refresh token is long lived

any subsequent access token request updates the cache.

access token generation uses a sign flag which is a set of random generated strings using werkzeug.security.secrets.token_urlsafe method
not quite sure if it is future proof as any subsquent request for an access token should produce a unique access token

( SHOULD LOOK INTO THIS IN THE NEAR FUTURE)

with the introduction of the token caching the application should be tested on an enviroment with a redis server configured

ALSO:

	each token ie access, refresh and activation token requires its own set of private key