User Registration and Login via the Interface
A complete user registration  basic factionality, the user can now register a new user account
but at the moment no application feedback to the user is implemented yet.
Once the notification component is implemented, user experience would be improved as detailed
notification would be displayed to the user showing the status of the application,

On chrome, since our api returns a 401 with a WWW-Authentication header set to Basic as the value
the chrome bromser launches a pop up propting the user user to enter their login credentials.

Still working on this!!!!