FrontEnd user Authentication and token handling logic setup
At the given moment a user can login to the account using basic authentication

authorization tokens are stored on the sessionStorage for smooth authorization
with the backend

As of now i have implemented access token refresh using  axios interceptors
In this case i don't need to know how long the access token need to live before it
expires

Some api code changes includes:

	1. Api sending static files via the email ( user confirmation link
is kinda hardcoded to the the source code. an idea to tackle this problem was
to allow the api to handle user account activation and password resets by
serving routes from the server side)

well this didn't go as planned as still i want to maintain the seperation of concerns .
such that the api does not serve any static files.