Database Service healthcheck set up The api service requires the database service to be fully initialized before any migrations are done
Up untill this point i have managed to write a health test to but will all the annoying warnings  from the
mysql server. attempts to sillence the warnings with --log-warnings=[level] makes the service initialization fail
as the database service is flagged as unhealthy.

Next Move to fix this: I'll probably use Postgresql server asit hase a bult in health checker.
:)